requirements/base.txt
python-slugify  # https://github.com/un33k/python-slugify
Pillow  # https://github.com/python-pillow/Pillow
argon2-cffi  # https://github.com/hynek/argon2_cffi
whitenoise  # https://github.com/evansd/whitenoise
redis  # https://github.com/redis/redis-py
hiredis  # https://github.com/redis/hiredis-py
uvicorn[standard]  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ  # https://github.com/joke2k/django-environ
django-model-utils  # https://github.com/jazzband/django-model-utils
django-allauth  # https://github.com/pennersr/django-allauth
django-crispy-forms  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework  # https://github.com/encode/django-rest-framework
django-cors-headers  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular  # https://github.com/tfranzel/drf-spectacular
django-htmx
django-treebeard
django-ninja
django-html_sanitizer
